requirements.txt
pip~=24.2
attrs~=23.2.0
Jinja2~=3.1.4
Flask>=2.0,<3.0
python-dotenv~=1.0.1
flask_jwt_extended~=4.4.4
flask_sqlalchemy~=3.0.5
pycountry~=22.3.5
dnspython~=2.1.0
Werkzeug~=3.0.3
pyodbc~=5.1.0
SQLAlchemy~=2.0.31
email_validator~=2.2.0
requests~=2.26.0